src/backend/requirements.txt
aiofiles
email_validator
jose
passlib
python_jose
PyYAML
fastapi
SQLAlchemy
aiosmtplib
uvicorn
gunicorn
python-multipart
aiosqlite
python-magic
python-magic-win64